We can show cards for English Wikipedia entries (vs rdf:seeAlso).
The card we can show depends on the linked URL being an English
Wikipedia page.  Sure, the plugin can check the predicate, but who
says this predicate will not be used for something else.  If the thing
we are referring to is a Wikipedia link, we probably want to render
the card.  This change reflects this logical reasoning.