We can render the hint card over the full referenced area
There is not that much to say to strip off the empty whitespace, it
seems.  That is to say, it makes sense when highlighting a card to
ensure the right area receives the href, but once that is set we
should follow the ideas put forward there.  Because we're not in a
steering position, this code can be removed.