Fix paths for dbpedia-query utility
We created the dbpedia-query utiltiy with a helper.  The generators
are great to help us get to the default case quickly.  Here we deviate
from the default case as we don't need the utility to be embedded in
the main application directly.  We can import it from the addon
instead.  As such we're moving the utility to a more normal place and
we're not pushing it to the consuming application anymore.

NOTE: we did not test if using a linked package somehow changes the
semantics of the addon vs app space with relative imports.  If after
publishing this addon something is broken, this change should be
revisited.