Remove generator function linter warning
It is our intention to support native execute functions which are not
tasks for cases like this.  Such use is not supported with native
classes yet.  As such we're ignoring the require-yield rule
temporarily.

There is still a warning for the use of _editor.  Not sure what the
best approach is for that.  We could search for a linter warning which
ignores variables starting with an _ as is common in TypeScript.